Reorganize RichFormattingMenu module directory structure
