+++ packages/slate-editor/src/modules/editor-v4-rich-formatting-menu/components/index.ts
@@ -1,0 +1,1 @@
+export { Toolbar } from './Toolbar';
