+++ packages/slate-editor/src/modules/editor-v4-rich-formatting-menu/lib/restoreSelection.ts
