--- packages/slate-editor/src/modules/editor-v4-rich-formatting-toolbar/index.ts
@@ -1,1 +1,0 @@
-export { RichFormattingMenu } from './components';
